always delete lib path
